Criando modulo para postagens.
